M duserissue

+++ src/main/java/com/CarRentProject/interfaces/CarService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface CarService {
     void saveCar(Car car);
     List<Car> getAllCars();
-    Car getCarById(Integer id);
+    Car getCarById(Long id);
     Car getCarByBrand(String brand);
     Car getCarByModel(String model);
     Car getCarByYearOfProduction(int yearOfProduction);
